tesis completada

+++ Front/dinnerSys-Front/src/pages/mesero/OpcionesMesero/PedidosMesero/PedidosMesero.jsx
@@ -2,12 +2,12 @@ import React, { useState, useEffect } from 'react';
 import { Table, InputGroup, FormControl, Button, Container, Alert } from 'react-bootstrap';
 import './PedidosMesero.css';
 import { getProducts } from '../../../../API/Productos';
-import { createOrder, getOrderXTableId } from '../../../../API/Pedidos';
+import { createOrder, getOrderXTableId, updateOrder } from '../../../../API/Pedidos';
 import { freeTable } from '../../../../API/Mesas';
-import { useAuth } from '../../../../auth/AuthProvider'; // Ajusta la ruta según la estructura de tu proyecto
+import { useAuth } from '../../../../auth/AuthProvider';
 
 function PedidosMesero({ mesa, pedido }) {
-  const { UserId } = useAuth(); // Obteniendo UserId del contexto de autenticación
+  const { UserId } = useAuth();
 
   const [filtros, setFiltros] = useState({
     nombre: '',
@@ -18,6 +18,7 @@ function PedidosMesero({ mesa, pedido }) {
   const [productosDisponibles, setProductosDisponibles] = useState([]);
   const [productosEnPedido, setProductosEnPedido] = useState([]);
   const [errorMensaje, setErrorMensaje] = useState('');
+  const [errorActualizarPedido, setErrorActualizarPedido] = useState(false); // Nuevo estado para controlar error de actualización
 
   useEffect(() => {
     const fetchData = async () => {
@@ -40,7 +41,7 @@ function PedidosMesero({ mesa, pedido }) {
     if (pedido && pedido.length > 0 && pedido[0].productos) {
       const productosExistentes = pedido[0].productos.map(p => ({
         ...p,
-        id: Date.now() + Math.random(), // Generar un ID único para cada producto
+        id: Date.now() + Math.random(),
         cantidad: p.cantidad || 1
       }));
       setProductosEnPedido(productosExistentes);
@@ -49,16 +50,45 @@ function PedidosMesero({ mesa, pedido }) {
   }, [pedido, mesa]);
 
   const agregarProducto = (producto) => {
-    const productoConID = { ...producto, id: Date.now(), cantidad: 1 };
-    const nuevosProductosEnPedido = [...productosEnPedido, productoConID];
-    setProductosEnPedido(nuevosProductosEnPedido);
-    localStorage.setItem(`productosEnPedidoMesa${mesa}`, JSON.stringify(nuevosProductosEnPedido));
+    const productoExistente = productosEnPedido.find(p => p.ProductoId === producto.ProductoId);
+    if (productoExistente) {
+      const nuevosProductosEnPedido = productosEnPedido.map(p =>
+        p.ProductoId === producto.ProductoId ? { ...p, cantidad: p.cantidad + 1 } : p
+      );
+      setProductosEnPedido(nuevosProductosEnPedido);
+      localStorage.setItem(`productosEnPedidoMesa${mesa}`, JSON.stringify(nuevosProductosEnPedido));
+    } else {
+      const productoConID = { ...producto, id: Date.now(), cantidad: 1 };
+      const nuevosProductosEnPedido = [...productosEnPedido, productoConID];
+      setProductosEnPedido(nuevosProductosEnPedido);
+      localStorage.setItem(`productosEnPedidoMesa${mesa}`, JSON.stringify(nuevosProductosEnPedido));
+    }
   };
 
-  const eliminarProducto = (id) => {
+  const eliminarProducto = async (id) => {
     const nuevosProductosEnPedido = productosEnPedido.filter(producto => producto.id !== id);
     setProductosEnPedido(nuevosProductosEnPedido);
     localStorage.setItem(`productosEnPedidoMesa${mesa}`, JSON.stringify(nuevosProductosEnPedido));
+
+    // Actualizar pedido en el servidor solo si ya se ha creado
+    if (pedido && pedido.length > 0) {
+      const pedidoActualizado = {
+        MeseroId: UserId,
+        MesaId: mesa,
+        lstProductos: nuevosProductosEnPedido.map(producto => ({ ProductoId: producto.ProductoId, Cantidad: producto.cantidad }))
+      };
+
+      try {
+        const isUpdate = await updateOrder(pedido[0].PedidoId, pedidoActualizado);
+        if (!isUpdate) {
+          setErrorActualizarPedido(true); // Mostrar error específico para la actualización
+        } else {
+          setErrorActualizarPedido(false); // Restablecer el estado de error de actualización
+        }
+      } catch (error) {
+        setErrorActualizarPedido(true); // Mostrar error específico para la actualización
+      }
+    }
   };
 
   const actualizarCantidadProducto = (id, cantidad) => {
@@ -72,7 +102,7 @@ function PedidosMesero({ mesa, pedido }) {
   const crearPedido = async () => {
     try {
       const newOrder = {
-        MeseroId: UserId, // Usando UserId del contexto de autenticación
+        MeseroId: UserId,
         MesaId: mesa,
         lstProductos: productosEnPedido.map(producto => ({ ProductoId: producto.ProductoId, Cantidad: producto.cantidad }))
       };
@@ -80,6 +110,7 @@ function PedidosMesero({ mesa, pedido }) {
       if (isCreated) {
         console.log("Pedido creado correctamente");
         setErrorMensaje('');
+        window.location.reload();
       } else {
         setErrorMensaje('Error al crear el pedido. Por favor, intenta de nuevo.');
       }
@@ -96,10 +127,6 @@ function PedidosMesero({ mesa, pedido }) {
     eliminarProducto(id);
   };
 
-  const handleGenerarPedido = () => {
-    // Lógica para generar el recibo del pedido
-  };
-
   const handleFiltrar = (e, campo) => {
     setFiltros({ ...filtros, [campo]: e.target.value });
   };
@@ -114,6 +141,7 @@ function PedidosMesero({ mesa, pedido }) {
     await freeTable(mesa, pedido[0].PedidoId);
     localStorage.removeItem(`productosEnPedidoMesa${mesa}`);
     setProductosEnPedido([]);
+    window.location.reload();
   };
 
   const calcularTotal = () => {
@@ -214,14 +242,14 @@ function PedidosMesero({ mesa, pedido }) {
                     <FormControl
                       type="number"
                       value={producto.cantidad}
-                      min="1"
                       onChange={(e) => actualizarCantidadProducto(producto.id, parseInt(e.target.value))}
+                      min="1"
                     />
                   </td>
                   <td>
                     <Button
                       variant="danger"
-                      className="listado-productos-button listado-productos-button-update btn-quitar"
+                      className="btn-quitar"
                       onClick={() => handleQuitar(producto.id)}
                     >
                       Quitar
@@ -232,31 +260,22 @@ function PedidosMesero({ mesa, pedido }) {
             </tbody>
           </Table>
         </div>
-
+        <div className="text-center mt-3">
+          <Button
+            variant="primary"
+            className="btn-liberar"
+            onClick={liberarMesa}
+          >
+            Liberar Mesa
+          </Button>
+        </div>
         <div className="total-container">
           <h3>Total: ${calcularTotal().toLocaleString('es-ES', {
             minimumFractionDigits: 0,
             maximumFractionDigits: 2  // Máximo 2 decimales
           })}</h3>
         </div>
-
-
-
-        <div className="btn-container">
-          <Button
-            variant="primary"
-            className="listado-productos-button listado-productos-button-update btn-generar"
-            onClick={liberarMesa}
-          >
-            Liberar Mesa
-          </Button>
-        </div>
-
-        {errorMensaje && (
-          <Alert variant="danger" className="text-center mt-3">
-            {errorMensaje}
-          </Alert>
-        )}
+        {errorMensaje && <Alert variant="danger" className="mt-3 text-center">{errorMensaje}</Alert>}
       </Container>
     </div>
   );
